Refactor code placement and name
