+++ common/src/main/java/com/wynntils/core/managers/ManagerRegistry.java
@@ -16,6 +16,7 @@ import com.wynntils.core.keybinds.KeyBindManager;
 import com.wynntils.core.webapi.WebManager;
 import com.wynntils.wynn.model.CharacterManager;
 import com.wynntils.wynn.model.WorldStateManager;
+import com.wynntils.wynn.model.container.ContainerQueryManager;
 import java.lang.reflect.InvocationTargetException;
 import java.util.ArrayList;
 import java.util.HashMap;
