+++ common/src/main/java/com/wynntils/wynn/model/container/ContainerContent.java
@@ -1,0 +1,12 @@
+/*
+ * Copyright © Wynntils 2022.
+ * This file is released under AGPLv3. See LICENSE for full license details.
+ */
+package com.wynntils.wynn.model.container;
+
+import java.util.List;
+import net.minecraft.network.chat.Component;
+import net.minecraft.world.inventory.MenuType;
+import net.minecraft.world.item.ItemStack;
+
+public record ContainerContent(List<ItemStack> items, Component title, MenuType menuType) {}
